chore: misc cleanup
